+++ packages/jaspr/lib/src/framework/framework.dart
@@ -1063,7 +1063,8 @@ abstract class Element implements BuildContext {
 
   void updateLastChild(Element? child) {
     _lastChild = child;
-    _lastRenderObjectElement = _lastChild?._lastRenderObjectElement ?? _lastChild?._prevSibling?._lastRenderObjectElement;
+    _lastRenderObjectElement =
+        _lastChild?._lastRenderObjectElement ?? _lastChild?._prevSibling?._lastRenderObjectElement;
     if (_parent?._lastChild == this && _parent?._lastRenderObjectElement != _lastRenderObjectElement) {
       _parent!.updateLastChild(this);
     }