+++ packages/jaspr/test/framework/elements/element_test.dart
@@ -6,7 +6,7 @@ import 'package:jaspr_test/jaspr_test.dart';
 void main() {
   group('elements', () {
     testComponents('correctly handles empty fragments', (tester) async {
-      final component = div( [
+      final component = div([
         text('A'),
         Fragment(children: [
           text('B'),
@@ -16,7 +16,8 @@ void main() {
       ]);
       tester.pumpComponent(component);
 
-      final render = find.byComponent(component).evaluate().first.lastRenderObjectElement?.renderObject as TestRenderObject;
+      final render =
+          find.byComponent(component).evaluate().first.lastRenderObjectElement?.renderObject as TestRenderObject;
       expect(render.tag, equals('div'));
       expect(render.children.length, equals(3));
       expect(render.children[0].text, equals('A'));