+++ packages/jaspr_lints/lib/src/lints/prefer_styles_getter.dart
@@ -9,8 +9,7 @@ class PreferStylesGetterLint extends DartLintRule {
       : super(
           code: LintCode(
             name: 'prefer_styles_getter',
-            problemMessage:
-                "Prefer using a getter over a variable declaration for style rules to support hot-reload.",
+            problemMessage: "Prefer using a getter over a variable declaration for style rules to support hot-reload.",
           ),
         );
 
@@ -18,22 +17,25 @@ class PreferStylesGetterLint extends DartLintRule {
   void run(CustomLintResolver resolver, ErrorReporter reporter, CustomLintContext context) {
     context.registry.addVariableDeclarationList((node) {
       if (node.variables.length != 1) return;
-      if (node.parent case TopLevelVariableDeclaration(metadata: var metadata, firstTokenAfterCommentAndMetadata: var token) || FieldDeclaration(metadata: var metadata, isStatic: true, firstTokenAfterCommentAndMetadata: var token) when metadata.any(checkCss)) {
+      if (node.parent
+          case TopLevelVariableDeclaration(metadata: var metadata, firstTokenAfterCommentAndMetadata: var token) ||
+              FieldDeclaration(metadata: var metadata, isStatic: true, firstTokenAfterCommentAndMetadata: var token)
+          when metadata.any(checkCss)) {
         final start = token.offset;
         final end = node.variables.first.name.end;
-        reporter.atOffset(offset: start, length: end-start, errorCode: code, data: node);
+        reporter.atOffset(offset: start, length: end - start, errorCode: code, data: node);
       }
     });
   }
 
   bool checkCss(Annotation a) {
-    return a.name.name == 'css' && a.name.staticElement?.librarySource?.uri.toString() == 'package:jaspr/src/foundation/styles/css.dart';
+    return a.name.name == 'css' &&
+        a.name.staticElement?.librarySource?.uri.toString() == 'package:jaspr/src/foundation/styles/css.dart';
   }
 
   @override
   List<Fix> getFixes() => [ReplaceWithGetterFix()];
 }
-
 
 class ReplaceWithGetterFix extends DartFix {
   @override
@@ -46,10 +48,9 @@ class ReplaceWithGetterFix extends DartFix {
   ) {
     if (analysisError.data case VariableDeclarationList node) {
       reporter.createChangeBuilder(message: 'Replace with getter', priority: 1).addDartFileEdit((builder) {
-
         final start = node.firstTokenAfterCommentAndMetadata.offset;
         final end = node.variables.first.equals!.end;
-        builder.addReplacement(SourceRange(start, end-start), (edit) {
+        builder.addReplacement(SourceRange(start, end - start), (edit) {
           edit.write(node.type?.toSource() ?? 'List<StyleRule>');
           edit.write(' get ');
           edit.write(node.variables.first.name.lexeme);
@@ -59,4 +60,3 @@ class ReplaceWithGetterFix extends DartFix {
     }
   }
 }
-