+++ packages/jaspr_content/CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased breaking
+
+- **Breaking**: Changed `MemoryPage.builder` to accept a `BuildContext` instead of a `Page` parameter.
+
 ## 0.1.1
 
 - `jaspr` upgraded to `0.19.0`