+++ packages/jaspr_content/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 - **Breaking**: Changed `MemoryPage.builder` to accept a `BuildContext` instead of a `Page` parameter.
 
+- Fixed typo with `keepSuffixPattern` parameter in `FilesystemLoader` and `GithubLoader`.
+- Adding or removing content files will now properly update the routes.
+
 ## 0.1.1
 
 - `jaspr` upgraded to `0.19.0`